Basic output for face shape data

Function performs some of the routine steps for getting landmark data ready for shape analyses. Data can be imported from a directory of Webmorph template files, or by reading an existing TPS file. For more details see vignette: vignette("intro", package = "facefuns")

basics(
  path_to_data,
  p = 189,
  remove_points = NA,
  rotate = c(NA, "flipX", "flipY", "rotateC", "rotateCC"),
  plot = TRUE,
  pc_criterion = "broken_stick"
)

Arguments

path_to_data
Path to .tps file or directory containing .tem files

p
Number of landmarks images were delineated with. Defaults to value for FRL standard template.

remove_points
If reading from template files, it can be specified whether any landmarks should be removed. If no value is entered, all landmarks are retained.

rotate
The type of rotation or flip to be performed. Specimens can be flipped with respect to x or y axes, or rotated clockwise or counter-clockwise. See geomorphs's rotate.coords

plot
Plot landmark coordinates of sample. See geomorphs's plotAllSpecimens

pc_criterion
Criterion used to choose which PCs to retain. See selectPCs

Value

Returns a list of the following:

data

Three-dimensional array containing original data

aligned

Three-dimensional array containing Procrustes-aligned/rotated data

scores

Principal component scores for each face

pcs

List of selected PCs, their importance and method used to select PCs

ref

Coordinates of sample average

plot

Coordinates of sample average at -3SD and +3SD for each selected PC