Quick start to shape analyses of 2-D data

Performs some of the routine steps for getting landmark data ready for shape analyses, such as Procrustes alignment and principal component analysis. For more details see vignette: vignette("intro", package = "facefuns")

facefuns2d(
  data,
  pc_criterion = "broken_stick",
  plot_sample = TRUE,
  auto_rotate = TRUE,
  quiet = FALSE
)

Arguments

data
Three-dimensional array of dimensions p, k, and n. p = number of landmarks, k = dimension (2D or 3D), n = number of specimens

pc_criterion
Criterion used to choose which PCs to retain. See select_pcs

plot_sample
Plot sample to check data. See plotAllSpecimens

auto_rotate
Landmark templates are sometimes no longer upright after Procrustes-alignment. Auto-rotate uses rotate.coords to guess which type of rotation is required

quiet
Print short summary of loaded data

Value

Returns a list of the following components:

array

Three-dimensional array containing Procrustes-aligned data

average

Coordinates of sample average for plotting

pc_info

List of selected PCs (including their SD, variance explained and cumulative variance explained), number of selected PCs, criterion used to select PCs

pc_scores

Principal component scores

pc_plot

PCs for plotting. Will by default create list of coordinates for all selected PCs at +/- 3SDs. To create plots of other PCs or at different level of SD, please see plot_2dpcs

summary

Short summary of key descriptives